Final Commit (Finished Game)
A short run down for personal use due to non-frequent commits.

Version 1: Game consists of a cube that can jump when space is pressed.
Half of the level completed, small tweaks need to be made to keep jump
height consistent.

Version 2: Level is completed. Player can now jump both when space
and/or touch is pressed. Geoplex - Solar Rain added as the song to be
played during the level. Fixed various bugs pertaining to jump height.

Version 3: Added new art for cubes and spikes as well as backgrounds for
the level and the main menu. Added an attempts counter and an options
and instructions menu. Options menu allows you to turn on vsync and
control the volume of the game. Changed jumping from adding force on the
Y velocity to straight increasing the Y velocity, making jumping more
consistent.

Version 4: Updated to unity 5.4 and created a Mac build to go alongside
the Windows and Android builds. Scrapped WebGL build due to it being
unstable. Polished the game, making UI look nice and making sure it
scales correctly no matter what device it is played on. Changed the way
the player checks to see if it is currently on the ground from using
colliders to using a Physics raycast. The raycast is much more accurate
and solves the jumping only slightly bug 8/10 times unless the jump
button is spammed, cannot fix this.

Final Version: Deleted objects and sprites that are no longer shaving
off around 2mb from the final file size, bring it to the max of 26mb.
Fixed pre-loading, allow the level to be loaded almost instantly as the
audio is now loaded on game launch. Made all of the code more readable
and maintainable and deleted code that isn't used often as well as any
debugging code that is no longer needed. Finally, created a build for
Windows, Mac and Android and tested all but Mac so that they work. Link
here: https://pjaer.itch.io/one-button-runner